Nearly don Power Point
Nearly finished the Power Point i will sent it to pears for a pear review and get feed back from them. i also tested it the size and made sure i didnt add to much by seing if i can see averthing clearly while editing it from my back on the seat of the airplane so around 30 cm which scaled up should be fine.  I also added a java file to do some testing for of how it could be implemented.